Avoid log of zero in loss calculations
Summary:
log(0) produces a Divide by 0 exception even though the result is
-Inf which gets converted to a finite value for HALF type. This
change adds the smallest denorm value for HALF and smallest non-denorm
values for FLOAT when computing the log of probability in cross
entropy loss calculation.

Test Plan: Extended tests to include a log(0) calculation

Reviewers: timp

Reviewed By: timp

Differential Revision: https://phabricator.sourcevertex.net/D11913

+++ tests/EncodingTest.cpp
@@ -61,7 +61,7 @@ copyIndices(const std::vector<std::uint64_t> &indices,
     std::int64_t min =
         -static_cast<std::int64_t>(std::numeric_limits<IndexType>::min());
     const auto range = max + min;
-    BOOST_CHECK(indices[i] <= range);
+    BOOST_CHECK(static_cast<int64_t>(indices[i]) <= range);
     typed[i] = static_cast<IndexType>(indices[i]);
   }
 }
